+++ rector.php
@@ -30,6 +30,8 @@ return static function (RectorConfig $rectorConfig): void {
         __DIR__ . '/_bffWeb/tests',
         __DIR__ . '/_bffApi/src',
         __DIR__ . '/_bffApi/tests',
+        __DIR__ . '/_metadata/src',
+        __DIR__ . '/_metadata/tests',
     ]);
 
     $rectorConfig->symfonyContainerXml(__DIR__ . '/var/cache/dev/App_KernelDevDebugContainer.xml');