+++ CHANGELOG.md
@@ -5,6 +5,45 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - XXXX-XX-XX
+### Added
+* Module metadata management [WIP]
+* Third Party APIs [WIP]
+
+## [v0.7.17] - 2024-08-23
+- ### Maintenance release
+- 
+## [v0.7.16] - 2024-08-23
+- ### Maintenance release
+
+## [v0.7.15] - 2024-08-22
+- ### Maintenance release
+
+## [v0.7.14] - 2024-08-20
+- ### Maintenance release
+
+## [v0.7.13] - 2024-08-18
+- ### Maintenance release
+
+## [v0.7.12] - 2024-08-18
+- ### Maintenance release
+
+## [v0.7.11] - 2024-08-18
+- ### Maintenance release
+
+## [v0.7.10] - 2024-08-17
+- ### Maintenance release
+
+## [v0.7.9] - 2024-08-17
+- ### Maintenance release
+
+## [v0.7.8] - 2024-08-17
+- ### Maintenance release
+
+## [v0.7.7] - 2024-08-17
+- ### Maintenance release
+
+## [v0.7.6] - 2024-08-17
+- ### Maintenance release
 
 ## [v0.7.5] - 2024-08-16
 - ### Added